#985 Move NativeProxyLookup to hartshorn-introspect
NativeProxyLookup being in hartshorn-proxy caused a cyclic dependency for hartshorn-introspect tests to hartshorn-proxy. This lookup is however not specific to Hartshorn proxies, and is therefore safe and logical to move to hartshorn-introspect
