+++ hartshorn-introspect/hartshorn-introspect-reflection/src/test/java/test/org/dockbox/hartshorn/util/introspect/ReflectionTypeParameterIntrospectionTests.java
@@ -16,7 +16,7 @@
 
 package test.org.dockbox.hartshorn.util.introspect;
 
-import org.dockbox.hartshorn.proxy.lookup.NativeProxyLookup;
+import org.dockbox.hartshorn.util.introspect.NativeProxyLookup;
 import org.dockbox.hartshorn.util.introspect.Introspector;
 import org.dockbox.hartshorn.util.introspect.annotations.VirtualHierarchyAnnotationLookup;
 import org.dockbox.hartshorn.util.introspect.reflect.ReflectionIntrospector;
